Default implementation of dnssec_summary()
